Moved mousebindings into seperate file+class
